% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reikna_hlutfall.R
\name{reikna_hlutfall}
\alias{reikna_hlutfall}
\title{Reikna hlutfoll}
\usage{
reikna_hlutfall(value, kvardi = 0:10, punktar = 1, prenta_lysandi = TRUE)
}
\arguments{
\item{kvardi}{bil sem kvardinn er a t.d. 0:10}

\item{punktar}{fjoldi punkkta a kvarda sem a ad taka saman, ef punktar = 1
tha er 1 punktur a moti punkti, ef punktar = 2 tha eru tveir punktar teknir
saman i einn t.d. 1, 2, 3, 4, 5, 6 yrdi 1-2, 2-3, 3-4, 5-6}
}
\value{
Skilar gagnaramma med tidni, hlutfalli, safntidni, maelitolum og
lysingu
}
\description{
Reikna hlutfoll
}
\examples{
Daemi ur bok Kolen og Brennan (2004) bls 40
value <- c(0,0,1,1,1,2,2,3,3,4)
reikna_hlutfall(value, kvardi = 0:4)
}
