% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reikna_hlutfall.R
\name{z_score}
\alias{z_score}
\title{Kvordun
Dæmi frá Kolen og Brennan (2004) bls 40
value <- c(0,0,1,1,1,2,2,3,3,4)}
\usage{
z_score(value)
}
\description{
Kvordun
Dæmi frá Kolen og Brennan (2004) bls 40
value <- c(0,0,1,1,1,2,2,3,3,4)
}
